use /play endpoint
